[ALPS05587429] wlan:Bridge dumpTXD to CHIP_DBG_OPS
[Detail]
Bridge dumpTXD to CHIP_DBG_OPS for connac2.x TXD debugging

[Solution]
Bridge dumpTXD to CHIP_DBG_OPS for connac2.x TXD debugging

MTK-Commit-Id: 3b7eebff6faaa445b7ae69e170ca5f8a1b8ee974

CR-Id: ALPS05587429
Feature: [Module]Wi-Fi Driver

+++ chips/common/dbg_connac.c
@@ -1208,7 +1208,7 @@ static char *q_idx_lmac_str[] = {"WMM0_AC0", "WMM0_AC1", "WMM0_AC2", "WMM0_AC3",
 	"Band1_ALTX", "Band1_BMC", "Band1_BNC", "Band1_PSMP",
 	"Invalid"};
 
-void halDumpTxdInfo(IN struct ADAPTER *prAdapter, uint32_t *tmac_info)
+void halDumpTxdInfo(IN struct ADAPTER *prAdapter, uint8_t *tmac_info)
 {
 	struct TMAC_TXD_S *txd_s;
 	struct TMAC_TXD_0 *txd_0;
